Math: allow angles outside [0, 360) as input for math_sin and math_cos

+++ include/math.h
@@ -73,10 +73,9 @@ INLINE i32 math_brad(i32 angle) {
     return angle * 0x2000 / 45;
 }
 
-// output in range [-0x4000, +0x4000]
-extern i32 math_sin(u16 angle);
+extern i32 math_sin(i32 angle);
 
-INLINE i32 math_cos(u16 angle) {
+INLINE i32 math_cos(i32 angle) {
     return math_sin(angle + math_brad(90));
 }
